// src/app/modules/articles/constants/article-action-names.enum.ts
export enum ArticleActionNames {
  GetAllArticles = 'Get All Articles',
  GetAllArticlesSuccess = 'Get All Articles Success',
  GetAllArticlesFailure = 'Get All Articles Failure',
  GetArticle = 'Get Article',
  GetArticleFailure = 'Get Article Failure',
  GetArticleSuccess = 'Get Article Success',
  FavoriteArticle = 'Favorite Article',
  FavoriteArticleFailure = 'Favorite Article Failure',
  FavoriteArticleSuccess = 'Favorite Article Success',
  UnfavoriteArticle = 'Unfavorite Article',
  UnfavoriteArticleFailure = 'Unfavorite Article Failure',
  UnfavoriteArticleSuccess = 'Unfavorite Article Success',
  AddCommentToArticle = 'Add Comment to Article',
  AddCommentToArticleFailure = 'Add Comment to Article Failure',
  AddCommentToArticleSuccess = 'Add Comment to Article Success',
  FollowProfile = 'Follow Profile',
  FollowProfileFailure = 'Follow Profile Failure',
  FollowProfileSuccess = 'Follow Profile Success',
  UnfollowProfile = 'Unfollow Profile',
  UnfollowProfileFailure = 'Unfollow Profile Failure',
  UnfollowProfileSuccess = 'Unfollow Profile Success',
  GetTags = 'Get Tags',
  GetTagsFailure = 'Get Tags Failure',
  GetTagsSuccess = 'Get Tags Success',
  GetFeed = 'Get Feed',
  GetFeedFailure = 'Get Feed Failure',
  GetFeedSuccess = 'Get Feed Success',
}